requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types @ file:///opt/miniconda3/conda-bld/annotated-types_1761744956819/work
anthropic @ file:///Users/ec2-user/croot/anthropic_1766054381478/work
anyio @ file:///Users/ec2-user/buildout/croot/anyio_1761936861605/work
attrs==25.4.0
babel==2.18.0
beautifulsoup4 @ file:///Users/ec2-user/croot/beautifulsoup4-split_1770291772119/work
certifi @ file:///Users/ec2-user/croot/certifi_1767659195195/work/certifi
cffi==2.0.0
charset-normalizer==3.4.4
contourpy @ file:///opt/miniconda3/conda-bld/contourpy_1763643920606/work
courlan==1.3.2
cryptography==46.0.5
cycler @ file:///Users/ec2-user/croot/cycler_1766067932347/work
dateparser==1.3.0
distro @ file:///Users/ec2-user/buildout/croot/distro_1759958175746/work
fonttools @ file:///Users/ec2-user/buildout/croot/fonttools_1768900114828/work
frozenlist==1.8.0
google-auth==2.48.0
google-genai==1.64.0
greenlet==3.3.2
h11 @ file:///opt/miniconda3/conda-bld/h11_1761931258926/work
html2text==2025.4.15
htmldate==1.9.4
httpcore @ file:///Users/ec2-user/buildout/croot/httpcore_1761937127030/work
httpx @ file:///Users/ec2-user/buildout/croot/httpx_1761937519393/work
idna @ file:///Users/ec2-user/buildout/croot/idna_1763763781171/work
jiter @ file:///opt/miniconda3/conda-bld/jiter_1762897313381/work
jusText==3.0.2
kiwisolver @ file:///opt/miniconda3/conda-bld/kiwisolver_1764159406929/work
lxml==6.0.2
lxml_html_clean==0.4.3
markdown-it-py @ file:///Users/ec2-user/croot/markdown-it-py_1767001622536/work
matplotlib==3.10.8
mdurl @ file:///Users/ec2-user/buildout/croot/mdurl_1759953843084/work
multidict==6.7.1
numpy @ file:///Users/ec2-user/croot/numpy_and_numpy_base_1770732576917/work/dist/numpy-2.4.2-cp314-cp314-macosx_12_0_arm64.whl#sha256=5a68f9bea9c38fa44c030e527959aa2d30d56d94400abc91fcb6ea3a754b6e38
openai @ file:///Users/ec2-user/croot/openai_1767167813200/work
outcome==1.3.0.post0
packaging @ file:///opt/miniconda3/conda-bld/packaging_1761049079068/work
pandas @ file:///Users/ec2-user/croot/pandas_1770642934667/work
pillow @ file:///Users/ec2-user/croot/pillow_1769433331593/work
playwright==1.58.0
playwright-stealth==2.0.2
propcache==0.4.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic @ file:///opt/miniconda3/conda-bld/pydantic_1764083526311/work
pydantic_core @ file:///opt/miniconda3/conda-bld/pydantic-core_1764009780495/work
pyee==13.0.1
Pygments @ file:///Users/ec2-user/buildout/croot/pygments_1763762315285/work
pyparsing @ file:///opt/miniconda3/conda-bld/pyparsing_1763973789564/work
PySocks==1.7.1
python-dateutil @ file:///Users/ec2-user/buildout/croot/python-dateutil_1759952060416/work
python-dotenv==1.2.1
python-telegram-bot==22.6
pytz==2025.2
regex==2026.2.19
requests==2.32.5
rich @ file:///opt/miniconda3/conda-bld/rich_1761820389859/work
rsa==4.9.1
setuptools==80.10.2
six @ file:///Users/ec2-user/buildout/croot/six_1759951952559/work
sniffio @ file:///opt/miniconda3/conda-bld/sniffio_1764329151061/work
sortedcontainers==2.4.0
soupsieve @ file:///Users/ec2-user/buildout/croot/soupsieve_1761936742519/work
tenacity==9.1.4
tld==0.13.1
tornado @ file:///Users/ec2-user/buildout/croot/tornado_1768896066734/work
tqdm @ file:///Users/ec2-user/croot/tqdm_1771398702970/work
trafilatura==2.0.0
trio==0.33.0
trio-websocket==0.12.2
typing-inspection @ file:///Users/ec2-user/buildout/croot/typing-inspection_1761937969604/work
typing_extensions @ file:///Users/ec2-user/buildout/croot/typing_extensions_1759952168048/work
tzlocal==5.3.1
undetected-chromedriver==3.5.5
urllib3==2.6.3
webdriver-manager==4.0.2
websocket-client==1.9.0
websockets==15.0.1
wheel==0.46.3
wsproto==1.3.2
yarl==1.22.0
